requirements.txt
streamlit
langchain
langchain-community
huggingface_hub
python-dotenv
langchain_openai
pymupdf
streamlit_pdf_viewer